# BACKEND/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BINAR CAR CHALLENGE CHAPTER 6 (OpenAPI 3.0)",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api/v1"
    }
  ],
  "tags": [
    {
      "name": "autentikasi",
      "description": "Login"
    },
    {
      "name": "Mobil",
      "description": "Operation"
    },
    {
      "name": "pengguna",
      "description": "Access"
    }
  ],
  "paths": {
    "/register": {
      "post": {
        "tags": [
          "autentikasi"
        ],
        "summary": "Register",
        "description": "Register",
        "operationId": "register",
        "security": [
          {}
        ],
        "requestBody": {
          "description": "Required fields [email, password, name]",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterCred"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "OK"
                    },
                    "data": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "If failed to retrive the data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultErrMsg"
                }
              }
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": [
          "autentikasi"
        ],
        "summary": "Login",
        "description": "Login",
        "operationId": "login",
        "security": [
          {}
        ],
        "requestBody": {
          "description": "Required fields [email, password]",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "admin@gmail.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "login successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/LoginCred"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultErrMsg"
                }
              }
            }
          },
          "401": {
            "description": "invalid email / password",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "FAIL"
                    },
                    "message": {
                      "type": "string",
                      "example": "Email or password is incorrect"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/whoAmi": {
      "get": {
        "tags": [
          "pengguna"
        ],
        "operationId": "whoAmI",
        "summary": "Get user",
        "description": "Return current user data based on bearer token \n\n Allowed status: superadmin, admin, member",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "OK"
                    },
                    "data": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized, bearer token invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedMess"
                }
              }
            }
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": [
          "pengguna"
        ],
        "summary": "Get regist users",
        "description": "Return list of all users  \n\n Allowed status: superadmin, admin",
        "operationId": "getUserList",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    },
                    "count": {
                      "type": "integer",
                      "format": "int64"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized, bearer token invalid or auth level below superadmin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedMess"
                }
              }
            }
          },
          "403": {
            "description": "not getting permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrNotAllowed"
                }
              }
            }
          },
          "default": {
            "description": "Default error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tambahadmin": {
      "post": {
        "tags": [
          "pengguna"
        ],
        "operationId": "createAdmin",
        "summary": "Add admin",
        "description": "Add new admin \n\n Allowed statua: superadmin",
        "requestBody": {
          "description": "Required fields [email, password, name]",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterCred"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "create user successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Admin"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized, bearer token invalid or auth level below superadmin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedMess"
                }
              }
            }
          },
          "403": {
            "description": "not getting permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrNotAllowed"
                }
              }
            }
          },
          "default": {
            "description": "Default error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultErrMsg"
                }
              }
            }
          }
        }
      }
    },
    "/add": {
      "post": {
        "tags": [
          "mobil"
        ],
        "summary": "Add car",
        "operationId": "addNewCar",
        "description": "Add new car data to database \n\n Allowed roles: superadmin, admin",
        "requestBody": {
          "description": "Required fields [name, price, size, image]",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "name",
                  "price",
                  "size",
                  "image"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Sample Car"
                  },
                  "price": {
                    "type": "integer",
                    "example": 3000000
                  },
                  "size": {
                    "type": "string",
                    "enum": [
                      "Small",
                      "Medium",
                      "Large"
                    ]
                  },
                  "image": {
                    "type": "string",
                    "example": "3i3mfrnwjernwjekrwk"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "OK"
                    },
                    "data": {
                      "$ref": "#/components/schemas/carRes"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized, bearer token invalid or auth level below admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedMess"
                }
              }
            }
          },
          "default": {
            "description": "Default error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultErrMsg"
                }
              }
            }
          }
        }
      }
    },
    "/mobil": {
      "get": {
        "tags": [
          "mobil"
        ],
        "summary": "Get cars",
        "operationId": "getCarList",
        "description": "Return data list car \n\n Allowed status: superadmin, admin",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "OK"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "post": {
                          "$ref": "#/components/schemas/mobil"
                        },
                        "count": {
                          "type": "integer",
                          "example": 1
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized, bearer token invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedMess"
                }
              }
            }
          },
          "default": {
            "description": "Default error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultErrMsg"
                }
              }
            }
          }
        }
      }
    },
    "/mobil/{id}": {
      "get": {
        "tags": [
          "mobil"
        ],
        "summary": "Get car by id",
        "operationId": "getCarData",
        "description": "Return car data based on the car ID \n\n Allowed roles: superadmin, admin",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "Numeric ID of the car to get"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "OK"
                    },
                    "data": {
                      "$ref": "#/components/schemas/carRes"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized, bearer token invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedMess"
                }
              }
            }
          },
          "404": {
            "description": "Not Found, car Id not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "OK"
                    },
                    "data": {
                      "type": "string",
                      "example": null
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Default error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultErrMsg"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "mobil"
        ],
        "summary": "Edit car",
        "operationId": "editCarData",
        "description": "Modify car data based on the car ID \n\n Allowed roles: superadmin, admin",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "Numeric ID of the car to edit"
          }
        ],
        "requestBody": {
          "description": "Allowed fields to update: [name, price, size, image]",
          "content": {
            "aplication/json": {
              "schema": {
                "$ref": "#/components/schemas/carRes"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "OK"
                    },
                    "data": {
                      "$ref": "#/components/schemas/carRes"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized, bearer token invalid or auth level below admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedMess"
                }
              }
            }
          },
          "404": {
            "description": "Not Found, car Id not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "failed"
                    },
                    "data": {
                      "type": "string",
                      "example": null
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Default error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultErrMsg"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "mobil"
        ],
        "summary": "Delete car",
        "operationId": "deleteCarData",
        "description": "Delete car data based on the car ID \n\n Allowed roles: superadmin, admin",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "Numeric ID of the car to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "OK"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "integer",
                        "example": 1
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized, bearer token invalid or auth level below admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedMess"
                }
              }
            }
          },
          "404": {
            "description": "Not Found, car Id not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "OK"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "integer",
                        "example": 0
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Default error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultErrMsg"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "UnauthorizedMess": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Unauthorized"
          }
        }
      },
      "carRes": {
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 4
          },
          "name": {
            "type": "string",
            "example": "toyota"
          },
          "price": {
            "type": "string",
            "example": 3000000000000000
          },
          "size": {
            "type": "string",
            "example": "Large"
          },
          "image": {
            "type": "string",
            "example": "DHKAJDGJASGDJD"
          },
          "cratedBy": {
            "type": "string",
            "example": "nama super admin"
          },
          "updatedBy": {
            "type": "string",
            "example": "nama super admin"
          },
          "deletedBy": {
            "type": "string",
            "example": "nama super admin"
          },
          "createdAt": {
            "type": "string",
            "example": "2022-10-12T13:25:21.995Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2022-10-12T13:25:21.995Z"
          }
        }
      },
      "mobil": {
        "type": "array",
        "items": {
          "properties": {
            "id": {
              "type": "integer",
              "format": "int64",
              "example": 4
            },
            "name": {
              "type": "string",
              "example": "toyota"
            },
            "price": {
              "type": "integer",
              "example": 3000000
            },
            "size": {
              "type": "string",
              "example": "Large"
            },
            "image": {
              "type": "string",
              "example": "toyota.png"
            },
            "cratedBy": {
              "type": "string",
              "example": "nama super admin"
            },
            "updatedBy": {
              "type": "string",
              "example": "nama super admin"
            },
            "deletedBy": {
              "type": "string",
              "example": "nama super admin"
            },
            "createdAt": {
              "type": "string",
              "example": "2022-10-12T13:25:21.995Z"
            },
            "updatedAt": {
              "type": "string",
              "example": "2022-10-12T13:25:21.995Z"
            }
          }
        }
      },
      "Admin": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 4
          },
          "email": {
            "type": "string",
            "example": "email@account.com"
          },
          "password": {
            "type": "string",
            "example": "jhskjfgjksf"
          },
          "name": {
            "type": "string",
            "example": "myName"
          },
          "status": {
            "type": "string",
            "example": "admin"
          },
          "cratedBy": {
            "type": "string",
            "example": "nama super admin"
          },
          "updatedBy": {
            "type": "string",
            "example": "nama super admin"
          },
          "createdAt": {
            "type": "string",
            "example": "2022-10-12T13:25:21.995Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2022-10-12T13:25:21.995Z"
          }
        }
      },
      "ErrNotAllowed": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "FAIL"
          },
          "message": {
            "type": "string",
            "example": "you are not allowed"
          }
        }
      },
      "DefaultErrMsg": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "FAIL"
          },
          "message": {
            "type": "string",
            "example": "Error message"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 4
          },
          "email": {
            "type": "string",
            "example": "octa@account.com"
          },
          "password": {
            "type": "string",
            "example": "password"
          },
          "name": {
            "type": "string",
            "example": "octa"
          },
          "status": {
            "type": "string",
            "example": "admin"
          },
          "cratedBy": {
            "type": "string",
            "example": "default"
          },
          "updatedBy": {
            "type": "string",
            "example": "default"
          },
          "createdAt": {
            "type": "string",
            "example": "2022-10-12T13:25:21.995Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2022-10-12T13:25:21.995Z"
          }
        }
      },
      "RegisterCred": {
        "type": "object",
        "required": [
          "email",
          "password",
          "name"
        ],
        "properties": {
          "email": {
            "type": "string",
            "example": "octa26@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "octavia"
          },
          "name": {
            "type": "string",
            "example": "octa"
          }
        }
      },
      "LoginCred": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 4
          },
          "email": {
            "type": "string",
            "example": "octa@account.com"
          },
          "password": {
            "type": "string",
            "example": "octavia"
          },
          "name": {
            "type": "string",
            "example": "octa"
          },
          "status": {
            "type": "string",
            "example": "admin"
          },
          "cratedBy": {
            "type": "string",
            "example": "default"
          },
          "updatedBy": {
            "type": "string",
            "example": "default"
          },
          "createdAt": {
            "type": "string",
            "example": "2022-10-12T13:25:21.995Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2022-10-12T13:25:21.995Z"
          },
          "token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MiwiZW1haWwiOiJ2aWFAZ21haWwuY29tIiwiaWF0IjoxNjY2NDA5ODQwfQ.dZ3lNBVo3dFWDfi14UtOHQWA65DT9dfiwsqYzHpq3pY"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}